Add money object to handle donation amount

+++ db/migrate/20131013093059_create_donations.rb
@@ -1,7 +1,7 @@
 class CreateDonations < ActiveRecord::Migration
   def change
     create_table :donations do |t|
-      t.integer :amount
+      t.money :amount
       t.string :name
       t.string :email
